+++ script/config/template.lua
@@ -1,5 +1,6 @@
 local util   = require 'utility'
 local define = require 'proto.define'
+local diag   = require 'proto.diagnostic'
 
 ---@class config.unit
 ---@field caller function
@@ -232,7 +233,7 @@ local template = {
                                             >> util.deepCopy(define.BuiltIn),
     ['Lua.diagnostics.enable']              = Type.Boolean >> true,
     ['Lua.diagnostics.globals']             = Type.Array(Type.String),
-    ['Lua.diagnostics.disable']             = Type.Array(Type.String),
+    ['Lua.diagnostics.disable']             = Type.Array(Type.String << util.getTableKeys(diag.getDiagAndErrNameMap())),
     ['Lua.diagnostics.severity']            = Type.Hash(
                                                 Type.String << util.getTableKeys(define.DiagnosticDefaultNeededFileStatus, true),
                                                 Type.String << {